requirements.txt
PlexAPI==4.13.4
python-dotenv==1.0.0
